Update patch set 1
Patch Set 1:

Steve, I have diff'd the I727 and I717 sources.  They are very similar.  The I717 might even be a bit new as I see some obvious bug fixes there that aren't in I727.  But then there are some diffs that are not clear which version to go with.

I was planning to submit a dependent patch that would bring I727 and I717 together.  Perhaps we can churn on that patch to get to a common code set?  (LMK if you want to go a different way)

As for T989, Mike is trying this kernel now.  We will know soon how compatible this is.  I bet it will work since the I727 prebuilt kernel works on T989.

Patch-set: 1