Update patch set 1
Patch Set 1:

The idea of the topology stuff is to tell the OS more details about how the CPUs in a device are arranged.

The idea is that in some architectures, it is more beneficial for power consumption to put all load on one core and let the other cores idle than to distribute it across multiple cores.

I know in the case of Samsung Exynos, this mechanism is already handled by the CPU hotplugging code.  In the event that both cores are active, it is better to distribute load evenly, because once you have more than one core active, cpuidle gets crippled and asymmetric loading will cause the frequency to ramp up.  As a result, some of these CPU topology patches can be quite detrimental to power consumption on those devices.

MSMs, as I understand them, support asymmetric clocking of each core, so the topology patches may have more benifit.  The thing is - those patches are really designed for more complex systems, such as when you have:
Multiple logical cores on a physical core (hyperthreading)
Multiple packages - e.g. scheduling gets far more interesting when you have two separate dual-core or quad-core CPUs.

In the case of dual-core-on-a-die CPUs with no hyperthreading, the benefit of the Linaro topology patches is questionable at best - and in fact on some architectures they behave detrimentally.

Patch-set: 1