Update patch set 1
Patch Set 1:

For whatever reason, don't know why yet, but with it we get sleep of death and reboots.  With the commit reverted, we get none of these issues.  We have spent the last 2 weeks trying different combinations and have narrowed it down to this commit.  It is frustrating that we don't have root cause yet, and we will go back to debug further.  But in the meantime, we need to revert this for stability.

Patch-set: 1