Update patch set 1
Patch Set 1:

Right now, no.  It works about 50% of the time and with this patch, the system seems more unstable, although it is hard to tell since we have other problems.  I'm going to take your advice on http://review.cyanogenmod.com/#/c/23370/ and debug the early suspend hooks.

Patch-set: 1