Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

Revised the patch to make it moar better. mdp4_overlay.c changes, basically moved the IFDEF up one line to encompass the condition. Didn't seem to make sense to make the contents of the condition conditional, and not the condition as well. 

This is cleaner, builds, and works just fine. Verified.

Patch-set: 2