Update patch set 1
Patch Set 1:

It can't be fixed unless we write a camera driver, because it talks to the ION device directly.

THere's no need for this patch right now though, if the setupBasePipe() issue can be figured out, it will work just fine with the old ION API. This is exactly what we do on D2.

Patch-set: 1