Update patch set 1
Patch Set 1:

We'll probably end up making it the default after a longer testing period, since it really does work well.

Also for future reference, there are a bunch of devices that share this kernel (with identical configurations) and also the base config in qcom-common. It's not likely that we would change something like this for just one device.

Patch-set: 1