Update patch set 1
Patch Set 1:

I evaluated this before, and it doesn't work quite as well as the stock mpdecision.  There is frequent, noticeable lag because it's not nearly as aggressive in onlining the second core. 

Really, it just needs some serious performance tuning. I would love to get rid of as many proprietary bits as possible.

Patch-set: 1