Update patch set 1
Patch Set 1:

Hi Steve,
the previous one you evaluated had my pyramid settings for on/offlining the 2nd core, these values in here should be more aggressive.
The array in question where this can be changed is:
static unsigned int NwNs_Threshold[8] = {12, 0, 25, 20, 32, 28, 0, 35};
old values: 	static unsigned int NwNs_Threshold[4] = {35, 0, 0, 5};
(which is about twice as aggressive as before)
Activate the debug switch if you want to adjust this to get a proper reading on the desired rq value and run adb shell cat /proc/kmsg | grep 'MPDEC'.

Explanation as to what does what inside the array:
This small formula calculates which value will be used (number_of_cpus_online - 1) * 2
The result of this formula will be the nwns_threshold where a new cpu is hotplugged.
The result of this formula + 1 will be the nwns_threshold where a cpu is unplugged.

nwns_threshold_x = runqueue threshold, if this is reached cpuX will be hot/unplugged
twts_threshold_x = time threshold, this amount of time must have passed for the related action to be taken (hot/unplug)

Example:
One cpu is online.
(1 - 1) * 2 = 0 ergo:
nwns_threshold_0 = cpu1 will be hotplugged at this value
((1 - 1) * 2) + 1 = 1
nwns_threshold_1 = cpu0 will be unplugged at this value
Since we can't unplug cpu0 this is '0'.

Two cpus are online.
(2 - 1) * 2 = 2 ergo:
nwns_threshold_2 = cpu2 will be hotplugged at this value
((2 - 1) * 2) + 1 = 3
nwns_threshold_3 = cpu1 will be unplugged at this value

etc...

The default values are:
NwNs_Threshold: 12, 0, 25, 20, 32, 28, 0, 35
TwTs_Threshold: 140, 0, 140, 190, 140, 190, 0, 190
Where the position and function of the number equals the result of the above explained formula.

(all times are in ms)

Patch-set: 1