Update patch set 8
Patch Set 8:

Sorry to be harsh with the -2, but:

This driver rewrite that I did for CM9 and CM10 breaks devices outside of I717, I727/T989 and T769/I577.  It also skips a lot of the workaround code that Samsung has for only-Samsung-knows issues with the hardware.  And because it is a rewrite, it breaks or complicates future updates from Samsung kernel releases. For these reasons, I do not want to forward port the driver to CM10.1.

I we must add BLN back in (and I know many users want it), I would prefer that we do it to the original driver.

Also, the defconfig change to enable BLN should be separate from the code change.  It makes future merges easier and some devices may not want to opt-in for BLN. See the 2nd bullet here: http://forum.xda-developers.com/showpost.php?p=21524480&postcount=125

Patch-set: 8