Update patch set 2
Patch Set 2: Code-Review-1

Turning the camera flash to "on" lights up the LED and keeps it on.  This seems to create an additional crashable bug in the gallery app as well as blowing up the Torch app.

This is probably a bad idea ...

Patch-set: 2
Reviewer: Dave Carey <6179@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1