Update patch set 2
Patch Set 2:

the early suspend and late resume does not really do anything in this driver and the ld9040 driver for herc/sr. the lcd is powered on and off by the board.It does not seem to matter what you change for this.  Might as well just remove it or #ifdef it out.

Patch-set: 2