Update patch set 1
Patch Set 1:

Matt - I'm waiting for Dferg to check this out as well.  Right now he's on vacation, something about making snow angels on a mountain and tumbling down face first.

Either way, with the two patches it does work on the skyrocket, however there is a 3 second lag between tapping record and it actually starting to record.  I haven't had a chance to logcat that yet to see what's going on, probably this weekend.

Patch-set: 1