Update patch set 1
Patch Set 1:

The Hercules and Skyrocket orientation sensors are rotated 90 degrees relative to each other, so a symptom of "upside down" probably can't be attributed to this orientation sensor difference.

Also, fixing a printk() %s issue (this patch) couldn't have caused the backwards and upside down issue. Look for something else that could be causing this.

Patch-set: 1