Update patch set 2
Patch Set 2: Code-Review+1

I don't have a Quincy to test, but the changes are in line with what's being done for the other msm8660 devices.

However, I'd recommend doing each of the variants' commits in a different repo branch, so that their Gerrit submissions don't depend on each other unnecessarily.

Patch-set: 2