Update patch set 2
Patch Set 2:

If CAF did that then I'm not sure... On the HTC Sensation (MSM8660) we tried getting writeback to work, but we were not successful (heck, an if statement in the code was even stopping us from using writeback). When the display HAL tried using the writeback fb, it spewed out errors and didn't work.

I think I'll stop here. The change has already been merged, so no reason to keep discussing it.

Patch-set: 2