Update patch set 1
Patch Set 1:

Samsung's ifdefs for this were an absolute mess, so I couldn't understand what was going on. This commit isn't intended to be used on its own, it's only meant to show the clean-up process (but even then, I didn't intend for it to not boot). If I mixed the clean-up together with the actual remapping then it would be hard to tell what I changed while remapping.

Would it be preferred that the device still boot with only this commit?

Patch-set: 1