Update patch set 4
Patch Set 4:

Final bump. If you're not interested in this then please abandon it. It's been ready for further review for months now. I'm not going to make any more amendments to my commit since this kernel is super broken and I don't want to spend time working on it.

Patch-set: 4