Update patch set 4
Patch Set 4: Code-Review-1

Sultan, thanks for the contribution, but I am not sure there is anyone on the CM team that wants to be responsible for a change of this scope (I'm not willing). As you mentioned, this is an old kernel on old devices.

Steve/Will/Dave, I'd be happy to remove my objection if one of you wants to handle any fallout.

Patch-set: 4
Label: Code-Review=-1